requirements.txt
playwright
fastapi
uvicorn
python-dotenv
google-generativeai
openai
chromadb
sentence-transformers
llama-cpp-python
jinja2
websockets
psutil
